# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sintade API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/recordings/{recording_id}/events": {
      "get": {
        "tags": [
          "recordings"
        ],
        "summary": "Live status of one of the caller's recordings.",
        "operationId": "recording_events",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "description": "The recording",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A `text/event-stream` of `status` events, each a RecordingStatus; a `gone` event ends it when the recording disappears",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingStatus"
                }
              }
            }
          },
          "401": {
            "description": "No valid session",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such recording in the caller's workspace",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/recordings/{recording_id}/links/{link_id}": {
      "delete": {
        "tags": [
          "sharing"
        ],
        "summary": "Revokes a link: it stops working immediately. Idempotent.",
        "operationId": "revoke_link",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "description": "The recording",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "link_id",
            "in": "path",
            "description": "The link",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Link revoked"
          },
          "401": {
            "description": "No valid session",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Missing CSRF token, or the role can't edit recordings",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such link in the caller's workspace",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "sharing"
        ],
        "summary": "Changes a live link's visibility, download flag or expiry.",
        "operationId": "update_link",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "description": "The recording",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "link_id",
            "in": "path",
            "description": "The link",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLinkBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Link updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareLinkResponse"
                }
              }
            }
          },
          "401": {
            "description": "No valid session",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Missing CSRF token, or the role can't edit recordings",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such live link in the caller's workspace",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid expiry",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/s/{slug}/events": {
      "get": {
        "tags": [
          "watch"
        ],
        "summary": "Live status of the recording behind a share link, for a viewer the link admits: the watch\npage waits on this while the recording is processing instead of polling.",
        "operationId": "events",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The link's slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A `text/event-stream` of `status` events, each a RecordingStatus",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingStatus"
                }
              }
            }
          },
          "401": {
            "description": "Sign in to watch this link",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such live link, or not for this viewer",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details, the body of every error response.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "RecordingStatus": {
        "type": "object",
        "description": "What a recording looks like right now.",
        "required": [
          "state",
          "hls",
          "sprite",
          "preview"
        ],
        "properties": {
          "hls": {
            "type": "boolean",
            "description": "The adaptive (HLS) ladder exists."
          },
          "preview": {
            "type": "boolean",
            "description": "The animated preview exists."
          },
          "sprite": {
            "type": "boolean",
            "description": "The scrub sprite and its `sprite.vtt` exist."
          },
          "state": {
            "type": "string",
            "description": "`uploading`, `processing`, `ready` or `failed`."
          }
        }
      },
...
      "ShareLinkResponse": {
        "type": "object",
        "required": [
          "id",
          "recording_id",
          "slug",
          "visibility",
          "allow_download",
          "created_at"
        ],
        "properties": {
          "allow_download": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "RFC 3339, or `null` for no expiry."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "recording_id": {
            "type": "string",
            "format": "uuid"
          },
          "revoked_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "RFC 3339, or `null` while the link works."
          },
          "slug": {
            "type": "string",
            "description": "The public identifier: the watch page is `/s/{slug}`."
          },
          "visibility": {
            "$ref": "#/components/schemas/VisibilityDto"
          }
        }
      },
...
      "UpdateLinkBody": {
        "type": "object",
        "properties": {
          "allow_download": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "A future RFC 3339 instant sets the expiry; `null` removes it; absent leaves it."
          },
          "visibility": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/VisibilityDto"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "VisibilityDto": {
        "type": "string",
        "enum": [
          "private",
          "workspace",
          "link",
          "public"
        ]
      },
...
    }
  }
}